Much more implemented for permission management

+++ backend/routes/admin/permissions/groups.py
@@ -16,13 +16,17 @@ from backend.models.permissions import (
     Permission,
     map_permissions_group,
 )
+from backend.models.user import User
+from backend.util.tokens import uses_token
+from backend.util.http_errors import BadRequest
 
 
 groups = Blueprint('groups', 'groups')
 
 
 @groups.post('/create')
-def create() -> IGroupId:
+@uses_token
+def create(user: User, *_) -> IGroupId:
     """
     Create a new permission group
 
@@ -44,19 +48,22 @@ def create() -> IGroupId:
     ## Returns:
     * `IGroupId`: ID for new group
     """
+    user.permissions.assert_can(Permission.ManagePermissionGroups)
     data = json.loads(request.data)
     name = data['name']
     permissions: list[IPermissionValueGroup] = data['permissions']
-    print(permissions)
+    if PermissionGroup.from_name(name) is not None:
+        raise BadRequest(f"A permission group with name {name} already exists")
     mapped_perms = map_permissions_group(permissions)
-    group = PermissionGroup.create(name, mapped_perms)
+    group = PermissionGroup.create(name, mapped_perms, False)
     return {
         "group_id": group.id
     }
 
 
 @groups.get('/list')
-def list_groups() -> IPermissionGroupList:
+@uses_token
+def list_groups(user: User, *_) -> IPermissionGroupList:
     """
     List available permission groups
 
@@ -79,6 +86,7 @@ def list_groups() -> IPermissionGroupList:
 
                             * `False`: permission denied
     """
+    user.permissions.assert_can(Permission.ManagePermissionGroups)
     groups = PermissionGroup.all()
     return {"groups": [
         {
@@ -97,7 +105,8 @@ def list_groups() -> IPermissionGroupList:
 
 
 @groups.put('/edit')
-def edit() -> dict:
+@uses_token
+def edit(user: User, *_) -> dict:
     """
     Edit an existing permission group
 
@@ -118,10 +127,20 @@ def edit() -> dict:
 
                     * `None`: permission inherited
     """
+    user.permissions.assert_can(Permission.ManagePermissionGroups)
     data = json.loads(request.data)
     group = PermissionGroup(PermissionGroupId(data['group_id']))
     name = data['name']
     permissions: list[IPermissionValueGroup] = data['permissions']
+
+    if (duplicate := PermissionGroup.from_name(name)) is not None:
+        # Only if it's not this group
+        if duplicate.id != group.id:
+            raise BadRequest(
+                f"A permission group with name {name} already exists")
+
+    if group.immutable:
+        raise BadRequest("Cannot edit immutable groups")
 
     group.name = name
     group.update_allowed(map_permissions_group(permissions))
@@ -129,14 +148,29 @@ def edit() -> dict:
 
 
 @groups.delete('/remove')
-def remove() -> dict:
+@uses_token
+def remove(user: User, *_) -> dict:
     """
     Remove an existing permission group
 
     ## Body:
     * `group_id` (`PermissionGroupId`): permission group ID
     """
-    data = json.loads(request.data)
-    group = PermissionGroup(PermissionGroupId(data['group_id']))
+    user.permissions.assert_can(Permission.ManagePermissionGroups)
+    group = PermissionGroup(PermissionGroupId(int(request.args['group_id'])))
+    transfer_group = PermissionGroup(
+        PermissionGroupId(int(request.args['transfer_group_id'])))
+
+    # Make sure we're not just transferring users to the same group
+    if group == transfer_group:
+        raise BadRequest(
+            "Cannot transfer users to same group when deleting a group")
+
+    if group.immutable:
+        raise BadRequest("Cannot remove immutable groups")
+
+    # Find users in group to delete, and transfer them to the transfer group
+    for u in group.get_users():
+        u.permissions.parent = transfer_group
     group.delete()
     return {}
